LayoutTests/fast/viewport/ios/shrink-to-fit-content-large-constant-width-expected.txt

This test verifies that a page with a constant width viewport bigger than the actual view width is scaled to fit the view.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS minScale is expectedScale
PASS innerWidth is expectedInnerWidth
PASS successfullyParsed is true

TEST COMPLETE
